src: polish code style
- Insert annotation on parameters
- Replace anonymous functions with lambdas
- Replace deprecated functions or APIs
- Remove some unused value or imports
- Polish redundant usages (casts, modifiers etc.)
- No dangling comment
- Extract more translatable strings
- Widely use data binding

+++ app/src/main/java/com/google/android/diskusage/Apps2SDLoader.java
@@ -1,4 +1,4 @@
-/**
+/*
  * DiskUsage - displays sdcard usage on android.
  * Copyright (C) 2008-2011 Ivan Volosyuk
  *
@@ -54,7 +54,7 @@ public class Apps2SDLoader {
   }
 
 
-  public FileSystemEntry[] load(final int blockSize) throws Throwable {
+  public FileSystemEntry[] load(final long blockSize) throws Throwable {
     UsageStatsManager usageStatsManager = (UsageStatsManager) diskUsage.getSystemService(Context.USAGE_STATS_SERVICE);
