+++ app/src/main/java/com/google/android/diskusage/entity/FileSystemEntrySmall.java
@@ -1,4 +1,4 @@
-/**
+/*
  * DiskUsage - displays sdcard usage on android.
  * Copyright (C) 2008-2011 Ivan Volosyuk
  *
@@ -37,7 +37,7 @@ public class FileSystemEntrySmall extends FileSystemEntry {
   }
 
   @Override
-  public FileSystemEntry filter(CharSequence pattern, int blockSize) {
+  public FileSystemEntry filter(CharSequence pattern, long blockSize) {
     return null;
   }
 }
