+++ app/src/main/java/com/google/android/diskusage/entity/FileSystemFreeSpace.java
@@ -1,4 +1,4 @@
-/**
+/*
  * DiskUsage - displays sdcard usage on android.
  * Copyright (C) 2008-2011 Ivan Volosyuk
  *
@@ -20,12 +20,12 @@
 package com.google.android.diskusage.entity;
 
 public class FileSystemFreeSpace extends FileSystemSpecial {
-  public FileSystemFreeSpace(String name, long size, int blockSize) {
+  public FileSystemFreeSpace(String name, long size, long blockSize) {
     super(name, size, blockSize);
   }
 
   @Override
-  public FileSystemEntry filter(CharSequence pattern, int blockSize) {
+  public FileSystemEntry filter(CharSequence pattern, long blockSize) {
     return null;
   }
 }
